# Fix #318 Automatic retry support with a good consideration of Slack API rate limites