--- a/slack-api-client/src/main/java/com/slack/api/methods/impl/AsyncExecutionSupplier.java
+++ b/slack-api-client/src/main/java/com/slack/api/methods/impl/AsyncExecutionSupplier.java
@@ -1,0 +1,16 @@
+package com.slack.api.methods.impl;
+
+import com.slack.api.methods.SlackApiException;
+import com.slack.api.methods.SlackApiResponse;
+
+import java.io.IOException;
+
+/**
+ * A Supplier that holds an API Method execution.
+ */
+@FunctionalInterface
+public interface AsyncExecutionSupplier<T extends SlackApiResponse> {
+
+    T execute() throws IOException, SlackApiException;
+
+}