--- a/slack-api-client/src/main/java/com/slack/api/methods/metrics/LastMinuteRequests.java
+++ b/slack-api-client/src/main/java/com/slack/api/methods/metrics/LastMinuteRequests.java
@@ -1,0 +1,6 @@
+package com.slack.api.methods.metrics;
+
+import java.util.concurrent.CopyOnWriteArrayList;
+
+public class LastMinuteRequests extends CopyOnWriteArrayList<Long> {
+}