--- a/slack-api-client/src/main/java/com/slack/api/methods/metrics/WaitingMessageIds.java
+++ b/slack-api-client/src/main/java/com/slack/api/methods/metrics/WaitingMessageIds.java
@@ -1,0 +1,6 @@
+package com.slack.api.methods.metrics;
+
+import java.util.concurrent.CopyOnWriteArrayList;
+
+public class WaitingMessageIds extends CopyOnWriteArrayList<String> {
+}