###  Summary

This pull request adds `AsyncMethodsClient` with rate-limiting support. With this Slack API client, you won't see rate-limited errors anymore.

The basic usage is fairly simple. It looks almost the same to the existing API client. But this one has a great consideration of Slack rate limits, so that it runs API calls more wisely.

```java
CompletableFuture<ChatPostMessageResponse> future =
  slack.methodsAsync(token).chatPostMessage(r -> r
    .channel("C1234567")
    .text("test"));
```

In addition, when it gets an HTTP 429 response from Slack API by any chance, [it automatically retries the request after the necessary time period to wait for the recovery](https://github.com/seratch/java-slack-sdk/blob/rate-limiter/slack-api-client/src/main/java/com/slack/api/methods/impl/AsyncRateLimitExecutor.java#L150-L155).

```java
} catch (SlackApiException e) {
  logSlackApiException(teamId, methodName, e);
  if (e.getResponse().code() == 429) {
    return enqueueThenWaitThenRun(messageId, teamId, methodName, params, methodsSupplier);
  }
  throw new MethodsCompletionException(null, e, null);
}
```

The core part of it is [AsyncRateLimitExecutor](https://github.com/seratch/java-slack-sdk/blob/rate-limiter/slack-api-client/src/main/java/com/slack/api/methods/impl/AsyncRateLimitExecutor.java). It internally runs `MethodsClient`'s API calls asynchronously as below ([the code](https://github.com/seratch/java-slack-sdk/blob/rate-limiter/slack-api-client/src/main/java/com/slack/api/methods/impl/AsyncRateLimitExecutor.java#L150-L155)).

```java
@Override
public CompletableFuture<ChatScheduleMessageResponse> chatScheduleMessage(ChatScheduleMessageRequest req) {
  return executor.execute(
    CHAT_SCHEDULE_MESSAGE, toMap(req), // these information will be used by the rate limiter
    () -> methods.chatScheduleMessage(req)  // a function that issues an HTTP request
  );
}
```

It works with [AsyncMethodsRateLimiter](https://github.com/seratch/java-slack-sdk/blob/rate-limiter/slack-api-client/src/main/java/com/slack/api/methods/impl/AsyncMethodsRateLimiter.java#L75-L96) that determines a safe timing to start API calls, particularly when the API client has already issued many requests in the last one minute.

```java
private WaitTime calculateWaitTime(String teamId, String key, int allowedRequests) {
  LastMinuteRequests lastMinuteRequests = metricsDatastore.getLastMinuteRequests(config.getExecutorName(), teamId, key);
  if (isBursty(lastMinuteRequests, allowedRequests)) {
    Double waitMillis = 180000D / allowedRequests; // change the pace (60 -> 180 seconds)
    return new WaitTime(waitMillis.longValue(), Pace.Bursty);
  } else if (isTooFastPaced(lastMinuteRequests, allowedRequests)) {
    Double waitMillis = 120000D / allowedRequests; // change the pace (60 -> 120 seconds)
    return new WaitTime(waitMillis.longValue(), Pace.TooFastPaced);
  } else if (isOptimalPace(lastMinuteRequests, allowedRequests)) {
    Double waitMillis = 60000D / allowedRequests; // keep the pace
    return new WaitTime(waitMillis.longValue(), Pace.Optimal);
  } else if (isBusy(lastMinuteRequests, allowedRequests)) {
    Double waitMillis = 30000D / allowedRequests; // starting make a bit slow
    return new WaitTime(waitMillis.longValue(), Pace.Safe);
  } else {
    return new WaitTime(0, Pace.Safe);
  }
}
```

All the rate-limiting tiers are defined like this: https://github.com/seratch/java-slack-sdk/blob/rate-limiter/slack-api-client/src/main/java/com/slack/api/methods/MethodsRateLimits.java

Also, newly introduced [`MethodsStats`](https://github.com/seratch/java-slack-sdk/blob/rate-limiter/slack-api-client/src/main/java/com/slack/api/methods/MethodsStats.java) shows the metrics on web API calls.

<details><summary>Show a stats example</summary>

```json
{
  "DEFAULT_SINGLETON_EXECUTOR": {
    "T1111111": {
      "all_completed_calls": {
        "auth.test": 2
      },
      "successful_calls": {
        "auth.test": 2
      },
      "unsuccessful_calls": {},
      "failed_calls": {},
      "current_queue_size": {
        "auth.test": 0
      },
      "last_minute_requests": {
        "auth.test": 2
      },
      "rate_limited_methods": {}
    },
    "T2222222": {
      "all_completed_calls": {
        "mpim.open": 1,
        "chat.postMessage": 113,
        "im.list": 1,
        "files.remote.update": 2,
        "chat.scheduledMessages.list": 4,
        "chat.delete": 6,
        "channels.setTopic": 1,
        "files.remote.info": 15,
        "mpim.history": 1,
        "search.messages": 2,
        "usergroups.disable": 1,
        "chat.deleteScheduledMessage": 3,
        "views.open": 2,
        "rtm.connect": 6,
        "groups.invite": 2,
        "files.remote.list": 2,
        "conversations.join": 1,
        "conversations.invite": 1,
        "reactions.get": 2,
        "groups.leave": 1,
        "reactions.remove": 2,
        "pins.add": 1,
        "channels.unarchive": 2,
        "groups.open": 1,
        "conversations.create": 11,
        "users.setPresence": 1,
        "im.close": 1,
        "channels.create": 1,
        "im.history": 1,
        "channels.replies": 2,
        "conversations.history": 4,
        "conversations.kick": 1,
        "channels.leave": 1,
        "chat.unfurl": 3,
        "conversations.open": 1,
        "stars.remove": 1,
        "groups.create": 1,
        "pins.list": 1,
        "files.upload": 7,
        "dnd.info": 1,
        "views.publish": 2,
        "conversations.close": 1,
        "reminders.delete": 2,
        "conversations.setTopic": 1,
        "reminders.complete": 2,
        "channels.mark": 1,
        "team.info": 5,
        "views.push": 2,
        "team.accessLogs": 2,
        "groups.replies": 1,
        "reminders.add": 2,
        "users.profile.get": 1,
        "stars.add": 1,
        "users.list": 18,
        "conversations.members": 6,
        "conversations.list": 49,
        "dnd.endSnooze": 2,
        "chat.getPermalink": 3,
        "team.profile.get": 2,
        "channels.info": 2,
        "chat.scheduleMessage": 3,
        "files.remote.share": 3,
        "users.info": 9,
        "files.remote.remove": 2,
        "groups.info": 1,
        "groups.archive": 2,
        "auth.test": 310,
        "groups.list": 2,
        "chat.meMessage": 1,
        "channels.setPurpose": 1,
        "usergroups.list": 3,
        "users.profile.set": 2,
        "im.mark": 2,
        "files.delete": 3,
        "conversations.leave": 1,
        "users.setActive": 1,
        "chat.update": 7,
        "users.conversations": 1,
        "mpim.mark": 2,
        "dialog.open": 1,
        "mpim.replies": 1,
        "usergroups.create": 2,
        "users.getPresence": 1,
        "rtm.start": 1,
        "search.files": 36,
        "dnd.endDnd": 1,
        "usergroups.users.update": 1,
        "pins.remove": 1,
        "files.list": 2,
        "files.remote.add": 3,
        "groups.setPurpose": 1,
        "channels.history": 5,
        "channels.list": 6,
        "chat.postEphemeral": 4,
        "conversations.info": 4,
        "team.billableInfo": 2,
        "files.sharedPublicURL": 2,
        "groups.mark": 1,
        "channels.archive": 2,
        "reactions.list": 2,
        "views.update": 2,
        "usergroups.update": 1,
        "emoji.list": 1,
        "users.identity": 1,
        "reminders.info": 2,
        "channels.rename": 1,
        "channels.kick": 1,
        "files.revokePublicURL": 2,
        "dnd.setSnooze": 1,
        "team.integrationLogs": 2,
        "usergroups.users.list": 2,
        "conversations.rename": 1,
        "im.replies": 1,
        "conversations.unarchive": 1,
        "groups.history": 1,
        "bots.info": 3,
        "search.all": 2,
        "conversations.archive": 10,
        "users.lookupByEmail": 1,
        "reactions.add": 2,
        "dnd.teamInfo": 1,
        "groups.createChild": 1,
        "conversations.replies": 2,
        "groups.rename": 1,
        "files.info": 3,
        "groups.close": 1,
        "stars.list": 2,
        "apps.uninstall": 1,
        "groups.kick": 1,
        "groups.setTopic": 1,
        "users.setPhoto": 1,
        "channels.invite": 3,
        "usergroups.enable": 1,
        "groups.unarchive": 1,
        "channels.join": 1,
        "mpim.close": 1,
        "users.deletePhoto": 1,
        "mpim.list": 1,
        "conversations.setPurpose": 1,
        "im.open": 1
      },
      "successful_calls": {
        "mpim.open": 1,
        "chat.postMessage": 112,
        "im.list": 1,
        "files.remote.update": 2,
        "chat.scheduledMessages.list": 4,
        "chat.delete": 6,
        "channels.setTopic": 1,
        "files.remote.info": 15,
        "mpim.history": 1,
        "usergroups.disable": 1,
        "search.messages": 2,
        "chat.deleteScheduledMessage": 3,
        "rtm.connect": 6,
        "groups.invite": 2,
        "files.remote.list": 2,
        "conversations.join": 1,
        "conversations.invite": 1,
        "reactions.get": 2,
        "reactions.remove": 2,
        "pins.add": 1,
        "channels.unarchive": 1,
        "groups.open": 1,
        "conversations.create": 9,
        "users.setPresence": 1,
        "im.close": 1,
        "channels.create": 1,
        "im.history": 1,
        "channels.replies": 2,
        "conversations.history": 4,
        "conversations.kick": 1,
        "channels.leave": 1,
        "chat.unfurl": 3,
        "conversations.open": 1,
        "stars.remove": 1,
        "groups.create": 1,
        "pins.list": 1,
        "files.upload": 7,
        "dnd.info": 1,
        "conversations.close": 1,
        "reminders.delete": 2,
        "conversations.setTopic": 1,
        "reminders.complete": 2,
        "channels.mark": 1,
        "team.info": 5,
        "team.accessLogs": 2,
        "groups.replies": 1,
        "reminders.add": 2,
        "users.profile.get": 1,
        "stars.add": 1,
        "users.list": 18,
        "conversations.members": 6,
        "conversations.list": 49,
        "dnd.endSnooze": 1,
        "chat.getPermalink": 3,
        "team.profile.get": 2,
        "channels.info": 2,
        "chat.scheduleMessage": 3,
        "files.remote.share": 3,
        "users.info": 9,
        "files.remote.remove": 1,
        "groups.info": 1,
        "groups.archive": 2,
        "auth.test": 310,
        "groups.list": 2,
        "chat.meMessage": 1,
        "channels.setPurpose": 1,
        "usergroups.list": 3,
        "users.profile.set": 2,
        "im.mark": 1,
        "files.delete": 3,
        "conversations.leave": 1,
        "users.setActive": 1,
        "chat.update": 7,
        "users.conversations": 1,
        "mpim.mark": 1,
        "mpim.replies": 1,
        "usergroups.create": 2,
        "users.getPresence": 1,
        "rtm.start": 1,
        "search.files": 36,
        "dnd.endDnd": 1,
        "usergroups.users.update": 1,
        "pins.remove": 1,
        "files.list": 2,
        "files.remote.add": 3,
        "groups.setPurpose": 1,
        "channels.history": 5,
        "channels.list": 6,
        "chat.postEphemeral": 4,
        "conversations.info": 4,
        "team.billableInfo": 2,
        "files.sharedPublicURL": 2,
        "groups.mark": 1,
        "channels.archive": 2,
        "reactions.list": 2,
        "usergroups.update": 1,
        "emoji.list": 1,
        "reminders.info": 2,
        "channels.rename": 1,
        "files.revokePublicURL": 2,
        "dnd.setSnooze": 1,
        "team.integrationLogs": 2,
        "usergroups.users.list": 1,
        "conversations.rename": 1,
        "im.replies": 1,
        "conversations.unarchive": 1,
        "groups.history": 1,
        "bots.info": 2,
        "search.all": 2,
        "conversations.archive": 10,
        "users.lookupByEmail": 1,
        "reactions.add": 2,
        "dnd.teamInfo": 1,
        "groups.createChild": 1,
        "conversations.replies": 2,
        "groups.rename": 1,
        "files.info": 3,
        "groups.close": 1,
        "stars.list": 2,
        "groups.kick": 1,
        "groups.setTopic": 1,
        "users.setPhoto": 1,
        "channels.invite": 2,
        "usergroups.enable": 1,
        "groups.unarchive": 1,
        "channels.join": 1,
        "mpim.close": 1,
        "users.deletePhoto": 1,
        "mpim.list": 1,
        "conversations.setPurpose": 1,
        "im.open": 1
      },
      "unsuccessful_calls": {
        "channels.kick": 1,
        "chat.postMessage": 1,
        "channels.unarchive": 1,
        "views.publish": 2,
        "channels.invite": 1,
        "mpim.mark": 1,
        "dialog.open": 1,
        "usergroups.users.list": 1,
        "conversations.create": 2,
        "views.open": 2,
        "dnd.endSnooze": 1,
        "files.remote.remove": 1,
        "views.update": 2,
        "views.push": 2,
        "groups.leave": 1,
        "users.identity": 1,
        "im.mark": 1,
        "apps.uninstall": 1
      },
      "failed_calls": {
        "bots.info": 1
      },
      "current_queue_size": {
        "chat.postMessage": 0,
        "views.publish": 0,
        "reminders.delete": 0,
        "search.messages": 0,
        "reminders.complete": 0,
        "views.open": 0,
        "bots.info": 0,
        "search.all": 0,
        "search.files": 0,
        "team.info": 0,
        "reactions.add": 0,
        "views.push": 0,
        "stars.list": 0,
        "team.accessLogs": 0,
        "reactions.get": 0,
        "team.billableInfo": 0,
        "reactions.remove": 0,
        "reminders.add": 0,
        "users.profile.get": 0,
        "users.list": 0,
        "conversations.list": 0,
        "chat.getPermalink": 0,
        "team.profile.get": 0,
        "reactions.list": 0,
        "users.info": 0,
        "views.update": 0,
        "auth.test": 0,
        "usergroups.list": 0,
        "users.profile.set": 0,
        "reminders.info": 0
      },
      "last_minute_requests": {
        "mpim.open": 1,
        "im.list": 1,
        "chat.postMessage_CT3DPH94K": 3,
        "chat.delete": 6,
        "chat.scheduledMessages.list": 4,
        "files.remote.update": 2,
        "chat.postMessage_CTEE3D2A0": 3,
        "channels.setTopic": 1,
        "files.remote.info": 15,
        "mpim.history": 1,
        "search.messages": 2,
        "usergroups.disable": 1,
        "chat.deleteScheduledMessage": 3,
        "rtm.connect": 6,
        "views.open": 2,
        "chat.postMessage_D1NA98CCV": 2,
        "chat.postMessage_CT3DPS5V1": 6,
        "groups.invite": 2,
        "files.remote.list": 2,
        "chat.postMessage_GTETG9916": 2,
        "conversations.join": 1,
        "conversations.invite": 1,
        "groups.leave": 1,
        "reactions.get": 2,
        "reactions.remove": 2,
        "pins.add": 1,
        "channels.unarchive": 2,
        "groups.open": 1,
        "conversations.create": 11,
        "users.setPresence": 1,
        "im.close": 1,
        "channels.create": 1,
        "chat.postMessage_#general": 1,
        "im.history": 1,
        "channels.replies": 2,
        "conversations.history": 4,
        "conversations.kick": 1,
        "channels.leave": 1,
        "chat.unfurl": 3,
        "conversations.open": 1,
        "stars.remove": 1,
        "chat.postMessage_CTGPDSQTZ": 2,
        "groups.create": 1,
        "pins.list": 1,
        "files.upload": 7,
        "dnd.info": 1,
        "views.publish": 2,
        "conversations.close": 1,
        "reminders.delete": 2,
        "conversations.setTopic": 1,
        "reminders.complete": 2,
        "channels.mark": 1,
        "team.info": 5,
        "views.push": 2,
        "team.accessLogs": 2,
        "groups.replies": 1,
        "reminders.add": 2,
        "users.profile.get": 1,
        "stars.add": 1,
        "chat.postMessage_#random": 69,
        "users.list": 18,
        "chat.postMessage_GHHT81A0N": 2,
        "conversations.members": 6,
        "chat.postMessage_random": 3,
        "conversations.list": 49,
        "dnd.endSnooze": 2,
        "chat.getPermalink": 3,
        "team.profile.get": 2,
        "channels.info": 2,
        "chat.scheduleMessage": 3,
        "files.remote.share": 3,
        "users.info": 9,
        "files.remote.remove": 2,
        "groups.info": 1,
        "groups.archive": 2,
        "auth.test": 310,
        "groups.list": 2,
        "chat.meMessage": 1,
        "channels.setPurpose": 1,
        "usergroups.list": 3,
        "users.profile.set": 2,
        "im.mark": 2,
        "files.delete": 3,
        "conversations.leave": 1,
        "users.setActive": 1,
        "chat.update": 7,
        "users.conversations": 1,
        "mpim.mark": 2,
        "dialog.open": 1,
        "mpim.replies": 1,
        "usergroups.create": 2,
        "users.getPresence": 1,
        "chat.postMessage_CLDASCNTB": 1,
        "rtm.start": 1,
        "search.files": 36,
        "dnd.endDnd": 1,
        "usergroups.users.update": 1,
        "pins.remove": 1,
        "files.list": 2,
        "files.remote.add": 3,
        "groups.setPurpose": 1,
        "channels.history": 5,
        "channels.list": 6,
        "chat.postEphemeral": 4,
        "conversations.info": 4,
        "team.billableInfo": 2,
        "chat.postMessage_C03E94MKS": 3,
        "chat.postMessage_C03E94MKU": 13,
        "files.sharedPublicURL": 2,
        "groups.mark": 1,
        "channels.archive": 2,
        "reactions.list": 2,
        "views.update": 2,
        "usergroups.update": 1,
        "emoji.list": 1,
        "users.identity": 1,
        "reminders.info": 2,
        "channels.kick": 1,
        "channels.rename": 1,
        "files.revokePublicURL": 2,
        "dnd.setSnooze": 1,
        "team.integrationLogs": 2,
        "usergroups.users.list": 2,
        "conversations.rename": 1,
        "im.replies": 1,
        "conversations.unarchive": 1,
        "groups.history": 1,
        "bots.info": 2,
        "search.all": 2,
        "conversations.archive": 10,
        "chat.postMessage_CTGPDPHGF": 3,
        "users.lookupByEmail": 1,
        "reactions.add": 2,
        "dnd.teamInfo": 1,
        "groups.createChild": 1,
        "conversations.replies": 2,
        "groups.rename": 1,
        "files.info": 3,
        "groups.close": 1,
        "stars.list": 2,
        "apps.uninstall": 1,
        "groups.kick": 1,
        "groups.setTopic": 1,
        "users.setPhoto": 1,
        "channels.invite": 3,
        "usergroups.enable": 1,
        "groups.unarchive": 1,
        "channels.join": 1,
        "mpim.close": 1,
        "users.deletePhoto": 1,
        "mpim.list": 1,
        "conversations.setPurpose": 1,
        "im.open": 1
      },
      "rate_limited_methods": {}
    }
  }
}
```
</details>

There are already two datastore implementations for the metrics/stats used by the rate limiter.

* [MetricsDatastore](https://github.com/seratch/java-slack-sdk/blob/rate-limiter/slack-api-client/src/main/java/com/slack/api/methods/metrics/MetricsDatastore.java) interface
* [MemoryMetricsDatastore](https://github.com/seratch/java-slack-sdk/blob/rate-limiter/slack-api-client/src/main/java/com/slack/api/methods/metrics/impl/MemoryMetricsDatastore.java) that works in-memory
* [RedisMetricsDatabase](https://github.com/seratch/java-slack-sdk/blob/rate-limiter/slack-api-client/src/main/java/com/slack/api/methods/metrics/impl/RedisMetricsDatastore.java) that works as a single datastore for multiple nodes

### Requirements (place an `x` in each `[ ]`)

…